requirements.txt
streamlit
python-pptx
pymupdf
requests
plotly
streamlit-authenticator